Move persistence components to a separate package, move missing components to synth and workload packages
