+++ csb/src/main/scala/edu/msstate/dasi/csb/persistence/SparkPersistence.scala
@@ -1,8 +1,9 @@
-package edu.msstate.dasi.csb
+package edu.msstate.dasi.csb.persistence
 
 import java.io.File
 
 import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
+import edu.msstate.dasi.csb.{Util, sc}
 import org.apache.hadoop.fs.FileUtil
 import org.apache.spark.graphx.{Edge, Graph, VertexId}
 import org.apache.spark.storage.StorageLevel
