+++ csb/src/main/scala/edu/msstate/dasi/csb/synth/kronMtx.scala
@@ -1,4 +1,4 @@
-package edu.msstate.dasi.csb
+package edu.msstate.dasi.csb.synth
 
 import scala.util.Random
