+++ csb/src/main/scala/edu/msstate/dasi/csb/workload/ClosenessCentrality.scala
@@ -1,4 +1,4 @@
-package edu.msstate.dasi.csb
+package edu.msstate.dasi.csb.workload
 
 import org.apache.spark.graphx._
